fixed the spelling in code

+++ pointers_arrays_strings/2-strncpy.c
@@ -8,7 +8,7 @@
  *
  * Return: pointer to the dest
  */
-char *_strncpy(char *dest, char *src, size_t n)
+char *_strncpy(char *dest, char *src, int n)
 {
 	size_t i;
